refactor: rename to async-generator module
